+++ src/realm/query_expression.hpp
@@ -251,6 +251,51 @@ struct Pow {
     typedef T type;
 };
 
+template <class T>
+struct Size {
+    int64_t operator()(T) const
+    {
+        return 1;
+    }
+    typedef T type;
+};
+
+template <>
+struct Size<StringData> {
+    int64_t operator()(StringData v) const
+    {
+        return v.size();
+    }
+    typedef StringData type;
+};
+
+template <>
+struct Size<BinaryData> {
+    int64_t operator()(BinaryData v) const
+    {
+        return v.size();
+    }
+    typedef BinaryData type;
+};
+
+template <>
+struct Size<ConstTableRef> {
+    int64_t operator()(ConstTableRef v) const
+    {
+        return v->size();
+    }
+    typedef ConstTableRef type;
+};
+
+template <>
+struct Size<ConstLinkViewRef> {
+    int64_t operator()(ConstLinkViewRef v) const
+    {
+        return v->size();
+    }
+    typedef ConstLinkViewRef type;
+};
+
 // Finds a common type for T1 and T2 according to C++ conversion/promotion in arithmetic (float + int => float, etc)
 template <class T1, class T2, bool T1_is_int = std::numeric_limits<T1>::is_integer || std::is_same<T1, null>::value,
           bool T2_is_int = std::numeric_limits<T2>::is_integer || std::is_same<T2, null>::value,
@@ -418,6 +463,8 @@ template <class oper, class TLeft = Subexpr, class TRight = Subexpr>
 class Operator;
 template <class oper, class TLeft = Subexpr>
 class UnaryOperator;
+template <class oper, class TLeft = Subexpr>
+class SizeOperator;
 template <class TCond, class T, class TLeft = Subexpr, class TRight = Subexpr>
 class Compare;
 template <bool has_links>
@@ -437,6 +484,8 @@ Query create(L left, const Subexpr2<R>& right)
 
 #ifdef REALM_OLDQUERY_FALLBACK // if not defined, then never fallback to query_engine.hpp; always use query_expression
     const Columns<R>* column = dynamic_cast<const Columns<R>*>(&right);
+    // TODO: recognize size operator expressions
+    // auto size_operator = dynamic_cast<const SizeOperator<Size<StringData>, Subexpr>*>(&right);
 
     if (column && ((std::numeric_limits<L>::is_integer && std::numeric_limits<R>::is_integer) ||
                    (std::is_same<L, double>::value && std::is_same<R, double>::value) ||
@@ -825,14 +874,13 @@ struct NullableVector {
     }
 
     template <typename Type = T>
-    typename std::enable_if<
-        realm::is_any<Type, float, double, OldDateTime, BinaryData, StringData, RowIndex, Timestamp, null>::value,
-        void>::type
+    typename std::enable_if<realm::is_any<Type, float, double, OldDateTime, BinaryData, StringData, RowIndex,
+                                          Timestamp, ConstTableRef, null>::value,
+                            void>::type
     set(size_t index, t_storage value)
     {
         m_first[index] = value;
     }
-
 
     inline util::Optional<T> get(size_t index) const
     {
@@ -1010,6 +1058,17 @@ inline void NullableVector<Timestamp>::set_null(size_t index)
     m_first[index] = Timestamp{};
 }
 
+// ConstTableRef
+template <>
+inline bool NullableVector<ConstTableRef>::is_null(size_t index) const
+{
+    return !bool(m_first[index]);
+}
+template <>
+inline void NullableVector<ConstTableRef>::set_null(size_t index)
+{
+    m_first[index].reset();
+}
 
 template <typename Operator>
 struct OperatorOptionalAdapter {
@@ -1554,7 +1613,6 @@ UnaryOperator<Pow<T>> power(const Subexpr2<T>& left)
     return {left.clone()};
 }
 
-
 // Classes used for LinkMap (see below).
 struct LinkMapFunction {
     // Your consume() method is given row index of the linked-to table as argument, and you must return whether or
@@ -1912,6 +1970,11 @@ public:
         return m_column->get_column_index();
     }
 
+    SizeOperator<Size<T>> size()
+    {
+        return SizeOperator<Size<T>>(this->clone(nullptr));
+    }
+
 private:
     // Column index of payload column of m_table
     mutable size_t m_column_ndx;
@@ -1996,7 +2059,6 @@ public:
         return string_compare<Like, LikeIns>(*this, col, case_sensitive);
     }
 };
-
 
 template <class T, class S, class I>
 Query string_compare(const Columns<StringData>& left, T right, bool case_sensitive)
@@ -2183,6 +2245,75 @@ private:
     LinkMap m_link_map;
 };
 
+template <class oper, class TLeft>
+class SizeOperator : public Subexpr2<Int> {
+public:
+    SizeOperator(std::unique_ptr<TLeft> left)
+        : m_left(std::move(left))
+    {
+    }
+
+    SizeOperator(const SizeOperator& other, QueryNodeHandoverPatches* patches)
+        : m_left(other.m_left->clone(patches))
+    {
+    }
+
+    // See comment in base class
+    void set_base_table(const Table* table) override
+    {
+        m_left->set_base_table(table);
+    }
+
+    void verify_column() const override
+    {
+        m_left->verify_column();
+    }
+
+    // Recursively fetch tables of columns in expression tree. Used when user first builds a stand-alone expression
+    // and binds it to a Query at a later time
+    const Table* get_base_table() const override
+    {
+        return m_left->get_base_table();
+    }
+
+    // destination = operator(left)
+    void evaluate(size_t index, ValueBase& destination) override
+    {
+        Value<Int>* d = dynamic_cast<Value<Int>*>(&destination);
+        REALM_ASSERT(d);
+
+        Value<T> left;
+        m_left->evaluate(index, left);
+
+        size_t sz = left.m_values;
+        d->init(left.m_from_link_list, sz);
+
+        for (size_t i = 0; i < sz; i++) {
+            auto elem = left.m_storage.get(i);
+            if (!elem) {
+                d->m_storage.set(i, util::none);
+            }
+            else {
+                d->m_storage.set(i, oper()(*elem));
+            }
+        }
+    }
+
+    std::unique_ptr<Subexpr> clone(QueryNodeHandoverPatches* patches) const override
+    {
+        return make_subexpr<SizeOperator>(*this, patches);
+    }
+
+    void apply_handover_patch(QueryNodeHandoverPatches& patches, Group& group) override
+    {
+        m_left->apply_handover_patch(patches, group);
+    }
+
+private:
+    typedef typename oper::type T;
+    std::unique_ptr<TLeft> m_left;
+};
+
 struct ConstantRowValueHandoverPatch : public QueryNodeHandoverPatch {
     std::unique_ptr<RowBaseHandoverPatch> row_patch;
 };
@@ -2279,6 +2410,11 @@ public:
         return LinkCount(m_link_map);
     }
 
+    LinkCount size() const
+    {
+        return LinkCount(m_link_map);
+    }
+
     template <typename C>
     SubColumns<C> column(size_t column_ndx) const
     {
@@ -2337,6 +2473,48 @@ private:
     friend class Table;
 };
 
+template <>
+class Columns<SubTable> : public Subexpr2<SubTable> {
+public:
+    Columns(const Columns& other, QueryNodeHandoverPatches* patches);
+
+    const Table* get_base_table() const override
+    {
+        return m_link_map.base_table();
+    }
+    void set_base_table(const Table* table) override
+    {
+        m_link_map.set_base_table(table);
+        m_column = &m_link_map.target_table()->get_column_table(m_column_ndx);
+    }
+
+    void verify_column() const override;
+
+    std::unique_ptr<Subexpr> clone(QueryNodeHandoverPatches* patches) const override
+    {
+        return make_subexpr<Columns<SubTable>>(*this, patches);
+    }
+
+    void evaluate(size_t index, ValueBase& destination) override;
+
+    SizeOperator<Size<ConstTableRef>> size()
+    {
+        return SizeOperator<Size<ConstTableRef>>(this->clone(nullptr));
+    }
+
+private:
+    Columns(size_t column_ndx, const Table* table, const std::vector<size_t>& links = {})
+        : m_link_map(table, links)
+        , m_column_ndx(column_ndx)
+        , m_column(&m_link_map.target_table()->get_column_table(column_ndx))
+    {
+    }
+
+    LinkMap m_link_map;
+    size_t m_column_ndx;
+    const SubtableColumn* m_column = nullptr;
+    friend class Table;
+};
 
 template <class Operator>
 Query compare(const Subexpr2<Link>& left, const ConstRow& row)
@@ -2573,15 +2751,7 @@ public:
     }
 
     // Load values from Column into destination
-    void evaluate(size_t index, ValueBase& destination) override
-    {
-        if (m_nullable && std::is_same<typename ColType::value_type, int64_t>::value) {
-            evaluate_internal<IntNullColumn>(index, destination);
-        }
-        else {
-            evaluate_internal<ColType>(index, destination);
-        }
-    }
+    void evaluate(size_t index, ValueBase& destination) override;
 
     bool links_exist() const
     {
@@ -2596,6 +2766,11 @@ public:
     size_t column_ndx() const noexcept
     {
         return m_sg ? get_column_base().get_column_index() : m_column_ndx;
+    }
+
+    SizeOperator<Size<T>> size()
+    {
+        return SizeOperator<Size<T>>(this->clone(nullptr));
     }
 
 private:
@@ -2619,6 +2794,18 @@ private:
             return *static_cast<SequentialGetter<ColType>&>(*m_sg).m_column;
     }
 };
+
+template <class T>
+void Columns<T>::evaluate(size_t index, ValueBase& destination)
+{
+    if (m_nullable && std::is_same<typename ColType::value_type, int64_t>::value) {
+        evaluate_internal<IntNullColumn>(index, destination);
+    }
+    else {
+        evaluate_internal<ColType>(index, destination);
+    }
+}
+
 
 template <typename T, typename Operation>
 class SubColumnAggregate;